+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add ability to create a transaction from a shared nonce
 * Implement JsonSerializable on Braintree Objects for easier logging (#200, thanks @hfmikep)
 * Fix spec to expect PayPal transaction to settle immediately after successful capture
+* Add `options` -> `paypal` -> `shipping` for creating & updating customers as well as creating payment methods
 
 ## 3.25.0
 * Add `submitForSettlement` option to `Subscription::retryCharge`